% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuego.R
\docType{package}
\name{fue}
\alias{fue}
\alias{fue-package}
\title{Run R functions from the CLI}
\description{
Rather than running R functions via a script, \code{fue::go} constructs a
function call and runs it directly, to be used via \code{R -e}.
}
\details{
Shell arguments that begin with "--" can specify names for arguments, everything
else is matched positionally using standard R matching rules.
}
\examples{
\dontrun{
  # the following all call scale(mtcars, center=TRUE, scale=FALSE)
  R -e "fue::go(scale)" --args  --no-scale --center --x=mtcars
  R -e "fue::go(scale)" --args  mtcars --scale=FALSE --center
  R -e "fue::go(scale)" --args  --no-scale mtcars TRUE
}

\dontrun{
  # Bash quoting can be tricky at times. This is a shell issue, not an R issue.
  R -e "fue::go(paste)" --args  '"Here I am"' '"Rock you like a Hurricane"'
}

\dontrun{
 # included is a shell script "fuego" that does all the boilerplate:
 fuego scale --no-scale mtcars TRUE
}


}
