% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuego.R
\name{scriptize}
\alias{scriptize}
\alias{print.script}
\title{Convience decorator for script functions}
\usage{
scriptize(FUN)

\method{print}{script}(x, ...)
}
\description{
This wraps a function, so that it evaluates when
printed; this removes the need for explicit parentheses.
}
\details{
*** HACK ***
}
